# web/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comic Reader API",
    "description": "Self-hostable comic reader. See comic-reader-spec.md.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/admin/email/status": {
      "get": {
        "tags": [
          "api::admin_email"
        ],
        "operationId": "status",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailStatusView"
                }
              }
            }
          },
          "403": {
            "description": "admin only"
          }
        }
      }
    },
    "/admin/email/test": {
      "post": {
        "tags": [
          "api::admin_email"
        ],
        "operationId": "test_send",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestEmailResp"
                }
              }
            }
          },
          "400": {
            "description": "no email on caller's account"
          },
          "403": {
            "description": "admin only"
          },
          "502": {
            "description": "transport error"
          }
        }
      }
    },
...
    "/filter-options/characters": {
      "get": {
        "tags": [
          "api::filter_options"
        ],
        "operationId": "characters",
        "parameters": [
          {
            "name": "library",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsView"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/account": {
      "patch": {
        "tags": [
          "api::account"
        ],
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResp"
                }
              }
            }
          },
          "400": {
            "description": "validation error"
          },
          "401": {
            "description": "current_password incorrect or session invalid"
          },
          "403": {
            "description": "OIDC users cannot edit email/password"
          },
          "409": {
            "description": "email already in use"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountReq": {
        "type": "object",
        "properties": {
          "confirm_password": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional second password field. The web form ships two `<input>`s\nfor confirmation; submitting them as `new_password` + `confirm_password`\nlets the no-JS fallback validate parity server-side. Ignored when\nabsent or equal."
          },
          "current_password": {
            "type": [
              "string",
              "null"
            ],
            "description": "Required when changing the password. Verified before any write."
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "description": "New email. Local users only — OIDC users cannot self-edit (the issuer\nowns this field)."
          },
          "new_password": {
            "type": [
              "string",
              "null"
            ],
            "description": "New password. Must be ≥ 12 chars per the local-auth policy."
          }
        }
      },
...
      "EmailStatusView": {
        "type": "object",
        "required": [
          "configured"
        ],
        "properties": {
          "configured": {
            "type": "boolean",
            "description": "`true` when the live sender is a real transport (Lettre or Mock),\n`false` when SMTP is unset and the Noop fallback is installed."
          },
          "last_duration_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "last_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_send_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_send_ok": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
      "MeResp": {
        "type": "object",
        "required": [
          "id",
          "display_name",
          "role",
          "csrf_token",
          "default_cover_solo",
          "activity_tracking_enabled",
          "timezone",
          "reading_min_active_ms",
          "reading_min_pages",
          "reading_idle_ms",
          "language",
          "exclude_from_aggregates",
          "show_marker_count"
        ],
        "properties": {
          "accent_color": {
            "type": [
              "string",
              "null"
            ],
            "description": "M4: accent palette token (e.g. `'amber'`, `'blue'`, `'emerald'`, `'rose'`)."
          },
          "activity_tracking_enabled": {
            "type": "boolean",
            "description": "M6a: per-user opt-out for reading-activity capture."
          },
          "csrf_token": {
            "type": "string"
          },
          "default_cover_solo": {
            "type": "boolean",
            "description": "Default for the reader's \"cover stands alone in double-page view\"\ntoggle. Per-series localStorage overrides at runtime."
          },
          "default_fit_mode": {
            "type": [
              "string",
              "null"
            ],
            "description": "M4: reader default fit mode — `'width' | 'height' | 'original'`."
          },
          "default_page_strip": {
            "type": "boolean",
            "description": "M4: when true the reader opens with the page strip visible."
          },
          "default_reading_direction": {
            "type": [
              "string",
              "null"
            ],
            "description": "Phase 3: per-user reader direction preference. `null` means \"auto\"\n(the reader still falls back to ComicInfo `Manga=YesAndRightToLeft`\ndetection per series)."
          },
          "default_view_mode": {
            "type": [
              "string",
              "null"
            ],
            "description": "M4: reader default view mode — `'single' | 'double' | 'webtoon'`."
          },
          "density": {
            "type": [
              "string",
              "null"
            ],
            "description": "M4: density token — `'comfortable' | 'compact'`."
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "exclude_from_aggregates": {
            "type": "boolean",
            "description": "Stats v2: opt-out from server-wide aggregates. When true, admin\ndashboards exclude this user's sessions from totals/top-series. Does\nnot affect personal `/me/reading-stats`."
          },
          "id": {
            "type": "string"
          },
          "keybinds": {
            "description": "M4: per-action keybind overrides for the reader. Empty object means\n\"use defaults\"."
          },
          "language": {
            "type": "string",
            "description": "Human-URLs M3: BCP-47 language tag used for the `NEXT_LOCALE` cookie\nand next-intl message bundle selection."
          },
          "reading_idle_ms": {
            "type": "integer",
            "format": "int32",
            "description": "M6a: idle threshold (ms) after which the client ends the session."
          },
          "reading_min_active_ms": {
            "type": "integer",
            "format": "int32",
            "description": "M6a: minimum accumulated active ms before a session is recorded."
          },
          "reading_min_pages": {
            "type": "integer",
            "format": "int32",
            "description": "M6a: minimum distinct pages before a session is recorded."
          },
          "role": {
            "type": "string"
          },
          "show_marker_count": {
            "type": "boolean",
            "description": "Markers M8: when true, /settings renders a count badge on the\nBookmarks sidebar row. Default false."
          },
          "theme": {
            "type": [
              "string",
              "null"
            ],
            "description": "M4: theme token — `'system' | 'dark' | 'light' | 'amber'`."
          },
          "timezone": {
            "type": "string",
            "description": "M6a: IANA timezone string for daily-bucket aggregations."
          }
        }
      },
...
      "OptionsView": {
        "type": "object",
        "required": [
          "values"
        ],
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TestEmailResp": {
        "type": "object",
        "required": [
          "delivered",
          "duration_ms",
          "to"
        ],
        "properties": {
          "delivered": {
            "type": "boolean"
          },
          "duration_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "to": {
            "type": "string"
          }
        }
      },
...
    }
  }
}